+++ editor/editor-core/src/__tests__/integration/card/inline-auth/inline-auth-try-another-account.ts
@@ -11,11 +11,15 @@ import {
 
 type ClientType = Parameters<typeof goToEditorTestingWDExample>[0];
 
+// FIXME: This test was automatically skipped due to failure on 13/04/2023: https://product-fabric.atlassian.net/browse/ED-17489
 BrowserTestCase(
   'inline: should open a new window to authenticate with a provider when connecting a different account',
 
   // Skipped test https://product-fabric.atlassian.net/browse/ED-17199
-  { skip: ['safari'] },
+  {
+    // skip: ['safari'],
+    skip: ['*'],
+  },
   async (client: ClientType) => {
     const page = await goToEditorTestingWDExample(client);
     const authorizationWindow = new AuthorizationWindow(client, page);