+++ engagement-platform/contextual-survey/src/components/FeedbackAcknowledgement.tsx
@@ -1,7 +1,8 @@
 /** @jsx jsx */
 import { css, jsx } from '@emotion/core';
 
-import { fontSize, gridSize } from '@atlaskit/theme/constants';
+import { fontSize } from '@atlaskit/theme/constants';
+import { token } from '@atlaskit/tokens';
 
 import SuccessContainer from './SuccessContainer';
 
@@ -12,7 +13,7 @@ export default () => (
         font-size: ${fontSize()}px;
         font-weight: 600;
         margin-top: 0;
-        line-height: ${gridSize() * 3}px;
+        line-height: ${token('font.lineHeight.300', '24px')};
       `}
     >
       Thanks for your feedback