+++ engagement-platform/contextual-survey/src/components/SurveyContainer.tsx
@@ -4,7 +4,7 @@ import { css, jsx } from '@emotion/core';
 import Button from '@atlaskit/button/custom-theme-button';
 import CrossIcon from '@atlaskit/icon/glyph/cross';
 import { N0, N50, N50A, N60A } from '@atlaskit/theme/colors';
-import { borderRadius, gridSize } from '@atlaskit/theme/constants';
+import { borderRadius } from '@atlaskit/theme/constants';
 import { token } from '@atlaskit/tokens';
 
 import { surveyInnerWidth } from '../constants';
@@ -14,15 +14,13 @@ interface Props {
   onDismiss: () => void;
 }
 
-const padding: number = gridSize() * 3;
-
 export default ({ children, onDismiss }: Props) => {
   return (
     <div
       css={css`
         background-color: ${token('elevation.surface.overlay', N0)};
         border-radius: ${borderRadius()}px;
-        padding: ${padding}px;
+        padding: ${token('space.300', '24px')};
         box-shadow: ${token(
           'elevation.shadow.overlay',
           `0 20px 32px -8px ${N50A}, 0 0 1px ${N60A}`,
@@ -33,8 +31,8 @@ export default ({ children, onDismiss }: Props) => {
       <div
         css={css`
           position: absolute;
-          top: ${padding - gridSize()}px;
-          right: ${padding - gridSize()}px;
+          top: ${token('space.200', '16px')};
+          right: ${token('space.200', '16px')};
         `}
       >
         <Button