+++ linking-platform/link-picker/src/ui/link-picker/link-search-list/link-search-no-results/no-results-svg/styled.ts
@@ -1,10 +1,10 @@
-import { gridSize as getGridSize } from '@atlaskit/theme/constants';
+import { token } from '@atlaskit/tokens';
 import { css } from '@emotion/react';
 
 const HEIGHT = 120;
 
 export const noResultsSVGStyles = css`
   height: ${HEIGHT}px;
-  margin: 0 auto ${getGridSize() * 3}px;
+  margin: 0 auto ${token('space.300', '24px')};
   display: block;
 `;