+++ linking-platform/link-picker/src/ui/link-picker/link-search-list/styled.ts
@@ -2,15 +2,13 @@ import { css } from '@emotion/react';
 // AFP-2532 TODO: Fix automatic suppressions below
 // eslint-disable-next-line @atlassian/tangerine/import/entry-points
 import { typography } from '@atlaskit/theme';
-import { gridSize as getGridSize } from '@atlaskit/theme/constants';
-
-const gridSize = getGridSize();
+import { token } from '@atlaskit/tokens';
 
 export const listContainerStyles = css`
   padding-top: 0;
   min-height: 80px;
-  margin-top: ${gridSize * 2}px;
-  margin-bottom: ${gridSize * 2}px;
+  margin-top: ${token('space.200', '16px')};
+  margin-bottom: ${token('space.200', '16px')};
   flex-grow: 1;
   display: flex;
   flex-direction: column;
@@ -35,5 +33,5 @@ export const listTitleStyles = css`
   ${typography.h100()}
   text-transform: uppercase;
   margin-top: 0;
-  margin-bottom: ${gridSize / 2}px;
+  margin-bottom: ${token('space.050', '4px')};
 `;