+++ linking-platform/link-picker/src/ui/link-picker/styled.ts
@@ -1,7 +1,6 @@
 import { token } from '@atlaskit/tokens';
 import { css } from '@emotion/react';
 import { N500 } from '@atlaskit/theme/colors';
-import { gridSize as getGridSize } from '@atlaskit/theme/constants';
 
 const LINK_PICKER_WIDTH_IN_PX = 342;
 
@@ -9,14 +8,13 @@ const LINK_PICKER_WIDTH_IN_PX = 342;
 const LINK_PICKER_MIN_HEIGHT_IN_PX_WITHOUT_DISPLAYTEXT = 142;
 const LINK_PICKER_MIN_HEIGHT_IN_PX_WITH_DISPLAYTEXT = 220;
 
-const gridSize = getGridSize();
-
 /**
  * Half padding on the top as the form field has a `gridSize()` margin top that cannot be overridden
  */
 export const rootContainerStyles = css`
   width: ${LINK_PICKER_WIDTH_IN_PX}px;
-  padding: ${gridSize}px ${gridSize * 2}px ${gridSize * 2}px;
+  padding: ${token('space.100', '8px')} ${token('space.200', '16px')}
+    ${token('space.200', '16px')};
   box-sizing: border-box;
   line-height: initial;
 `;
@@ -42,15 +40,15 @@ export const rootContainerStylesForLoaderWithoutDisplaytext = css`
 `;
 
 export const searchIconStyles = css`
-  margin-left: ${gridSize / 2}px;
+  margin-left: ${token('space.050', '4px')};
   color: ${token('color.icon', N500)};
   cursor: default;
 `;
 
 export const tabsWrapperStyles = css`
-  margin-top: ${gridSize * 1.5}px;
-  margin-left: -${gridSize}px;
-  margin-right: -${gridSize}px;
+  margin-top: ${token('space.150', '12px')};
+  margin-left: calc(-1 * ${token('space.100', '8px')});
+  margin-right: calc(-1 * ${token('space.100', '8px')});
 `;
 
 export const flexColumnStyles = css`
@@ -59,5 +57,5 @@ export const flexColumnStyles = css`
 `;
 
 export const formFooterMargin = css`
-  margin-top: ${gridSize * 2}px;
+  margin-top: ${token('space.200', '16px')};
 `;