+++ linking-platform/link-picker/src/ui/link-picker/text-input/styled.ts
@@ -1,11 +1,10 @@
 import { token } from '@atlaskit/tokens';
 import { css } from '@emotion/react';
 import { N500 } from '@atlaskit/theme/colors';
-import { gridSize } from '@atlaskit/theme/constants';
 
 export const clearTextButtonStyles = css`
   padding: 0;
-  margin-right: ${gridSize() / 2}px;
+  margin-right: ${token('space.050', '4px')};
   color: ${token('color.icon.subtle', N500)};
   background: transparent;
   border: none;
@@ -14,6 +13,6 @@ export const clearTextButtonStyles = css`
 
 export const fieldStyles = css`
   & + & {
-    margin-top: ${gridSize() * 2}px;
+    margin-top: ${token('space.200', '16px')};
   }
 `;